+++ allauth/socialaccount/providers/oauth2/client.py
@@ -45,7 +45,7 @@ class OAuth2Client(object):
         params.update(extra_params)
         return "%s?%s" % (authorization_url, urlencode(params))
 
-    def get_access_token(self, code):
+    def get_access_token(self, code, pkce_code_verifier=None):
         data = {
             "redirect_uri": self.callback_url,
             "grant_type": "authorization_code",
@@ -67,6 +67,8 @@ class OAuth2Client(object):
         if self.access_token_method == "GET":
             params = data
             data = None
+        if data and pkce_code_verifier:
+            data["code_verifier"] = pkce_code_verifier
         # TODO: Proper exception handling
         resp = requests.request(
             self.access_token_method,